ffmpeg workerpool fix
