+++ src/scraper/api/handlers/register.go
@@ -16,14 +16,17 @@ func RegisterHandler(r *mux.Router, svc core.Service) {
 	r.Handle("/song/{id}/", coreHandler.DownloadSong())
 	r.Handle("/playlist/{id}/", coreHandler.DownloadPlaylist())
 	r.Handle("/album/{id}/", coreHandler.DownloadAlbum())
+	//r.Handle("/show/{id}/", coreHandler.DownloadShow())
 
 	// fetch just the metadata
 	r.Handle("/meta/song/{id}/", coreHandler.ViewSongMeta())
 	r.Handle("/meta/playlist/{id}/", coreHandler.ViewPlaylistMeta())
 	r.Handle("/meta/album/{id}/", coreHandler.ViewAlbumMeta())
+	//r.Handle("/meta/show/{id}/", coreHandler.ViewShowMeta())
 
 	// status info
 	r.Handle("/status/song/{id}/", coreHandler.ViewSongProgress())
 	r.Handle("/status/playlist/{id}/", coreHandler.ViewPlaylistProgress())
 	r.Handle("/status/album/{id}/", coreHandler.ViewAlbumProgress())
+	//r.Handle("/status/show/{id}/", coreHandler.ViewShowProgress())
 }
